backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
torch
torchvision
pillow==10.4.0
